- added safer 64-bit support

+++ src/zipaccess.h
@@ -25,8 +25,8 @@ namespace hltypes
 		void close(Resource* resource, void* archive);
 		void* fopen(void* archivefile, const String& filename);
 		void fclose(void* file);
-		long fread(void* file, void* buffer, int count);
-		long fsize(void* archivefile, const String& filename);
+		int32_t fread(void* file, void* buffer, int32_t count);
+		int64_t fsize(void* archivefile, const String& filename);
 		void* freopen(void* file, void* archivefile, const String& filename);
 		Array<String> getFiles(void* archivefile);
 		FileInfo finfo(void* archivefile, const String& filename);
